Fixed issue with uncached birthday kings causing bot crash
On bot startup not all guild members are cached, becouse of that when attempting to give the birthday role to a guildMember the member might not exist in the cache due to not having any activity since last startup. This results in returning undefined insted of the requested guildMember (when searching by id) and will rise an unhandled execption as undefined has no property roles.

To fix that the function handeling birthday updates was added a fetch for all guild members before attempting to make changes to them.

	modified:   index.js