Update .dockerignore to include more file types
The .dockerignore file has been updated to include all .mp3, .wav, and .db files in any directory. This ensures that these files are not included when building the Docker image. Additionally, the .dockerignore file ignores itself, along with the .github/ directory, Dockerfile, and any .md files, are now also ignored. This update helps to improve the efficiency, cleanliness of the Docker build process and reduce image size.